build: Move stylesheet processing into pkg/lib/build-common.js to enable reusing
There are some quite delibarate hacks there, it worths to export these
so that other projects can keep in sync with cockpit more easily.
